backend/requirements.txt
Flask
Flask-CORS
pytesseract
Pillow
googletrans==4.0.0-rc1
gTTS
gunicorn
langdetect